[test](Nereids): add test for SimplifyComparisonPredicate
